+++ src/GitVersionCore.Tests/CommitDateTests.cs
@@ -4,7 +4,6 @@ using GitVersion;
 using GitVersion.Configuration;
 using GitVersion.Extensions;
 using GitVersion.VersionCalculation;
-using GitVersion.VersionFilters;
 using GitVersionCore.Tests.Helpers;
 using NUnit.Framework;
