+++ src/GitVersionCore.Tests/GitVersionExecutorTests.cs
@@ -3,9 +3,9 @@ using System.IO;
 using System.Text;
 using GitTools.Testing;
 using GitVersion;
-using GitVersion.Cache;
 using GitVersion.Configuration;
 using GitVersion.Logging;
+using GitVersion.VersionCalculation.Cache;
 using GitVersionCore.Tests.Helpers;
 using LibGit2Sharp;
 using Microsoft.Extensions.DependencyInjection;
