+++ src/GitVersionCore/Configuration/EffectiveConfiguration.cs
@@ -1,7 +1,6 @@
 using System.Collections.Generic;
 using GitVersion.Extensions;
 using GitVersion.VersionCalculation;
-using GitVersion.VersionFilters;
 
 namespace GitVersion.Configuration
 {
