+++ src/GitVersionCore/Configuration/IgnoreConfig.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using GitVersion.VersionFilters;
+using GitVersion.VersionCalculation;
 using YamlDotNet.Serialization;
 
 namespace GitVersion.Configuration
