+++ src/GitVersionCore/Helpers/GitRepositoryHelper.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using System.Linq;
-using GitVersion.Exceptions;
 using GitVersion.Extensions;
 using GitVersion.Logging;
 using LibGit2Sharp;
