+++ src/GitVersionCore/Model/Exceptions/GitVersionException.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace GitVersion.Exceptions
+namespace GitVersion
 {
     [Serializable]
     public class GitVersionException : GitToolsException
