+++ src/GitVersionCore/Model/Exceptions/WarningException.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace GitVersion.Exceptions
+namespace GitVersion
 {
     [Serializable]
     public class WarningException : Exception
