+++ src/GitVersionCore/Model/GitVersionCacheKey.cs
