+++ src/GitVersionCore/VersionCalculation/Cache/IGitVersionCacheKeyFactory.cs
@@ -1,0 +1,10 @@
+﻿using GitVersion.Cache;
+using GitVersion.Configuration;
+
+namespace GitVersion.VersionCalculation.Cache
+{
+    public interface IGitVersionCacheKeyFactory
+    {
+        GitVersionCacheKey Create(Config overrideConfig);
+    }
+}
