+++ src/GitVersionCore/VersionCalculation/SemanticVersioning/SemanticVersion.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Text.RegularExpressions;
-using GitVersion.Exceptions;
 
 namespace GitVersion
 {
