+++ src/GitVersionCore/VersionCalculation/SemanticVersioning/SemanticVersionBuildMetaData.cs
