+++ src/GitVersionCore/VersionCalculation/SemanticVersioning/SemanticVersionExtensions.cs
