+++ src/GitVersionCore/VersionCalculation/SemanticVersioning/SemanticVersionFormatValues.cs
