+++ src/GitVersionExe/ArgumentParser.cs
@@ -4,7 +4,6 @@ using System.Collections.Specialized;
 using System.IO;
 using System.Linq;
 using GitVersion.Configuration;
-using GitVersion.Exceptions;
 using GitVersion.Extensions;
 using GitVersion.Logging;
 using GitVersion.Model;
