+++ src/GitVersionTask/FileHelper.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
-using GitVersion.Exceptions;
 using Microsoft.Build.Framework;
 
 namespace GitVersion.MSBuildTask
